+++ js/dataset/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/dataset
 
+## 0.1.38
+
+### Dependency updates
+
+- @gradio/utils@0.4.2
+- @gradio/atoms@0.7.4
+- @gradio/upload@0.10.5
+- @gradio/client@0.19.4
+
 ## 0.1.37
 
 ### Dependency updates